Add in-progress WiGLE records for Albemarle County
Here are approximately half of the available WiGLE records for Albemarle County. (Strictly speaking these are not _Albemarle_ records, but the records for the quadrangle that includes Albemarle County, excluding all records are specifically flagged by WiGLE as being in Charlottesville. Which is to say there will be a few odd ones for outlying counties etc., e.g. Grottoes.) Toward #7.